@@ -1,8 +1,10 @@
 {
-  "name": "screeps-typescript-starter",
-  "version": "3.0.0",
+  "name": "screeps-bot-ts",
+  "version": "0.2.0",
   "description": "",
+  "private": true,
   "main": "index.js",
+  "type": "module",
   "//": "If you add or change the names of destinations in screeps.json, make sure you update these scripts to reflect the changes",
   "scripts": {
     "lint": "eslint \"src/**/*.ts\"",
@@ -19,22 +21,12 @@
     "watch-season": "rollup -cw --environment DEST:season",
     "watch-sim": "rollup -cw --environment DEST:sim"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/screepers/screeps-typescript-starter.git"
-  },
-  "author": "",
-  "license": "Unlicense",
-  "bugs": {
-    "url": "https://github.com/screepers/screeps-typescript-starter/issues"
-  },
-  "homepage": "https://github.com/screepers/screeps-typescript-starter#readme",
-  "engines": {
-    "node": "10.x || 12.x"
-  },
+  "repository": "private",
+  "author": "Serge Ivanov <lutvuk@gmail.com>",
+  "license": "MIT",
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^20.0.0",
-    "@rollup/plugin-node-resolve": "^13.0.4",
+    "@rollup/plugin-commonjs": "^25.0.7",
+    "@rollup/plugin-node-resolve": "^15.2.3",
     "@types/chai": "^4.1.6",
     "@types/lodash": "3.10.2",
     "@types/mocha": "^5.2.5",
@@ -42,27 +34,28 @@
     "@types/screeps": "^3.2.3",
     "@types/sinon": "^5.0.5",
     "@types/sinon-chai": "^3.2.0",
-    "@typescript-eslint/eslint-plugin": "^4.29.1",
-    "@typescript-eslint/parser": "^4.29.1",
-    "@typescript-eslint/typescript-estree": "^4.29.1",
+    "@typescript-eslint/eslint-plugin": "^7.0.2",
+    "@typescript-eslint/parser": "^7.0.2",
+    "@typescript-eslint/typescript-estree": "^7.0.2",
     "chai": "^4.2.0",
-    "eslint": "^7.32.0",
+    "eslint": "^8.56.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-import-resolver-typescript": "^2.4.0",
     "eslint-plugin-import": "^2.24.0",
     "eslint-plugin-prettier": "^3.4.0",
-    "lodash": "^3.10.1",
-    "mocha": "^5.2.0",
+    "lodash": "^4.17.21",
+    "mocha": "^10.3.0",
     "prettier": "^2.3.2",
-    "rollup": "^2.56.2",
+    "rollup": "^4.12.0",
     "rollup-plugin-clear": "^2.0.7",
     "rollup-plugin-screeps": "^1.0.1",
-    "rollup-plugin-typescript2": "^0.31.0",
+    "rollup-plugin-typescript2": "^0.36.0",
     "sinon": "^6.3.5",
     "sinon-chai": "^3.2.0",
     "ts-node": "^10.2.0",
     "tsconfig-paths": "^3.10.1",
-    "typescript": "^4.3.5"
+    "typescript": "^5.3.3",
+    "typescript-eslint": "^7.0.2"
   },
   "dependencies": {
     "source-map": "~0.6.1"
