@@ -1,9 +1,9 @@
-import {Dictionary} from "lodash";
+import { Dictionary } from "lodash";
 
 export const limits: Dictionary<number> = {
-    builder:   2,
-    carrier:   3,
-    claimer:   2,
-    harvester: 2,
-    upgrader:  2
+  builder: 2,
+  carrier: 3,
+  claimer: 2,
+  harvester: 2,
+  upgrader: 2
 };
